Initial import of original Hugo website
This is the initial import of the Hugo website, https://gohugo.io as it
was on the night of May 10th, 2024. In order to save space, as well as
there not being any plan to import future commits from the upstream
website, the original repository has been flattened (history restarted)
into this single commit.